Made a few changes to make it work on windows, some bug fixes as well as some other things, the code to convert the text to tokens is a lot less optimised now but its just more maintainable which is preferable for the meantime

+++ interpreter_common/memory.c
@@ -1,7 +1,9 @@
 #include "memory.h"
 
+#include <stddef.h>
+
 #define IMPLEMENT_GROWTH_FUNCTION(name, uppercase) \
-  inline int get_next_capacity_##name(int capacity) { \
+  inline size_t get_next_capacity_##name(size_t capacity) { \
     return capacity < GROWTH_MIN_##uppercase \
       ? GROWTH_MIN_##uppercase : capacity * GROWTH_RATE_##uppercase; \
   }
